requirements.txt
streamlit
pandas
Pillow
transformers
torch
torchvision
